refactor: Create Prompts module and App class
